requirements.txt
Flask == 1.0.2
gunicorn == 19.9.0
tensorflow == 1.13.1
numpy == 1.15.4
opencv-python == 4.1.0.25
keras == 2.2.4
scikit-learn == 0.20.3
Pillow == 6.0.0
